Moved Heater and Valve into Effectors.c
